+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added command export, import and clone functions to myems-api
 - added energy flow diagram export, import and clone functions to myems-api
 - added tariff export, import and clone functions to myems-api
+- added working calendar export, import and clone functions to myems-api
 ### Changed
 ### Fixed
 - fixed unkonwn meter uuid of energy flow diagram links in demo database