+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added virtual power plant export, import and clone functions to myems-api
 - added wind farm export, import and clone functions to myems-api
 - added rule export, import and clone functions to myems-api
+- added photovoltaic power station export, import and clone functions to myems-api
 ### Changed
 ### Fixed
 - fixed unkonwn meter uuid of energy flow diagram links in demo database